# Conflicts:
#	Assets/Art/3D/Gameplay/Dragons/dragon_jawfrey/Disguises/Skins/dragon_jawfrey_2_DIFF.tga
#	Assets/Art/3D/Gameplay/Dragons/dragon_jawfrey/Disguises/Skins/dragon_jawfrey_2_ingame_DIFF.tga
#	Assets/Game/Gameplay/Dragons/DragonMotion.cs
#	Assets/Resources/Game/Equipable/Skins/dragon_jawfrey/dragon_jawfrey_2_wings.mat
#	Assets/Resources/Rules/dragonDefinitions.xml


Former-commit-id: 503eec7a8085650e2b6654329d84f8213c27da84

+++ Assets/Game/Model/MissionsAndEvents/Trackers/TrackerFireRush.cs
@@ -58,7 +58,7 @@ public class TrackerFireRush : TrackerBase {
 	/// <param name="_tid">Description TID to be formatted.</param>
 	/// <param name="_targetValue">Target value. Will be placed at the %U0 replacement slot.</param>
 	/// <param name="_replacements">Other optional replacements, starting at %U1.</param>
-	override public string FormatDescription(string _tid, float _targetValue, params string[] _replacements) {
+	override public string FormatDescription(string _tid, long _targetValue, params string[] _replacements) {
 		// Singular/Plural issue (https://mdc-tomcat-jira100.ubisoft.org/jira/browse/HDK-1202)
 		// Figure out which tid to use
 		string timeTid = "TID_GEN_TIME";
